+++ index.html
@@ -14,7 +14,7 @@
     <script src="api-settings.js" defer></script>
     <script src="pomodoro.js" defer></script>
     <script src="eisenhower.js" defer></script>
-    <script src="day-planner.js" defer></script>
+    <script type="module" src="day-planner.js"></script>
     <script src="task-manager.js" defer></script>
     <script src="task-breakdown.js" defer></script>
     <script src="habit-tracker.js" defer></script>
